requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
alembic==1.13.0
sqlalchemy==2.0.23
passlib==1.7.4
bcrypt==4.1.1
pydantic==2.5.0
python-jose==3.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
sqlalchemy-utils==0.41.1